fix: prevent tools service overload with sitemap generation (#84)
* fix(sitemap): lower max pages and duration for sitemap crawler

* perf(sitemap): refactor callback to send a single testing request
This prevents unwanted (and unnecessary) load on the testing servers & database when first scanning the sitemap.
When scanning the sitemap, we might just find hundreds or thousands of URLs: sending an individual testing request
for each is likely to overload the servers and database. The service already has support for an array of pages,
so we just switch to that instead.